frontend: Split ApartmentDetailsPage into components
- Over 700 lines is a bit too much for a single file :)
- No changes to code
