+++ frontend/src/features/apartment/ApartmentDetailsPage/components/index.ts
@@ -1,0 +1,11 @@
+import ApartmentConditionsOfSaleCard from "./ApartmentConditionsOfSaleCard";
+import ApartmentMaximumPricesCard from "./ApartmentMaximumPricesCard";
+import ApartmentSalesPageLinkButton from "./ApartmentSalesPageLinkButton";
+import RemoveApartmentLatestSaleModalButton from "./RemoveApartmentLatestSaleModalButton";
+
+export {
+    ApartmentConditionsOfSaleCard,
+    ApartmentMaximumPricesCard,
+    ApartmentSalesPageLinkButton,
+    RemoveApartmentLatestSaleModalButton,
+};
