+++ frontend/src/features/apartment/ApartmentDetailsPage/index.ts
@@ -1,0 +1,3 @@
+import ApartmentDetailsPage from "./ApartmentDetailsPage";
+
+export default ApartmentDetailsPage;
